@@ -1,14 +1,13 @@
-﻿namespace CK.Core
+namespace CK.Core;
+
+/// <summary>
+/// Injectable reusable <see cref="UserMessageCollector"/> available as a Scoped service.
+/// </summary>
+public class ScopedUserMessageCollector : UserMessageCollector, IScopedAutoService
 {
-    /// <summary>
-    /// Injectable reusable <see cref="UserMessageCollector"/> available as a Scoped service.
-    /// </summary>
-    public class ScopedUserMessageCollector : UserMessageCollector, IScopedAutoService
+    /// <inheritdoc />
+    public ScopedUserMessageCollector( CurrentCultureInfo culture )
+        : base( culture )
     {
-        /// <inheritdoc />
-        public ScopedUserMessageCollector( CurrentCultureInfo culture )
-            : base( culture )
-        {
-        }
     }
 }
